requirements.txt
openai
python-dotenv
python-pptx
python-docx
ollama
websockets
pyautogui
pyperclip
websocket-client
pyaudio
cozepy